refactor(AMS-172): split resources.ts into enums,configs and utils for modularity
