+++ src/constants/resources/resourceEnums.ts
@@ -1,0 +1,116 @@
+// Resource routes enum
+export enum RESOURCE_ROUTES {
+  USERS = "users",
+  DASHBOARD = "dashboard",
+  DETAIL_DEVICE = "detail-device",
+  CHECKIN_CHECKOUT = "checkin_checkout",
+  ASSETS = "assets",
+  ASSETS_ASSIGN = "assets-assign",
+  ASSETS_READY_TO_DEPLOY = "assets-readyToDeploy",
+  ASSETS_PENDING = "assets-pending",
+  ASSETS_BROKEN = "assets-broken",
+  ASSETS_MAINTENANCE = "assets-maintenance",
+  TOOLS_ALL = "tools-all",
+  TOOLS_WAITING = "tools-waiting",
+  LICENSES = "licenses",
+  USERS_TOOLS = "users-tools",
+  USERS_TAX_TOKENS = "users-tax-tokens",
+  USERS_LICENSES = "users-licenses",
+  MODELS = "models",
+  CATEGORY = "category",
+  MANUFACTURES = "manufactures",
+  SUPPLIERS = "suppliers",
+  DEPARTMENT = "department",
+  LOCATION = "location",
+  TAX_TOKEN_ASSIGN = "tax_token_assign",
+  REQUEST = "request",
+  REPORT = "report",
+  MANAGER_USER = "manager_user",
+  ASSETS_WAITING_CONFIRM = "assets-waiting-confirm",
+  CONSUMABLES = "consumables",
+  CONSUMABLES_MAINTENANCE = "consumables-maintenance",
+  ACCESSORY = "accessory",
+  ACCESSORY_DETAILS = "accessory_details",
+  CONSUMABLE_DETAILS = "consumable_details",
+  SUPPLIER_DETAILS = "supplier_details",
+  LOCATION_DETAILS = "location_details",
+  MANUFACTURES_DETAILS = "manufactures_details",
+  WEBHOOK_DETAILS = "webhook_details",
+  ASSETS_EXPIRES = "assets-expires",
+  TAX_TOKEN = "tax_token",
+  TAX_TOKEN_WAITING = "tax_token_waiting",
+  CLIENT_ASSETS = "client-assets",
+  CLIENT_ASSET_ASSIGNED = "client-asset-assigned",
+  CLIENT_ASSET_READY_TO_DEPLOY = "client-asset-readyToDeploy",
+  CLIENT_ASSET_PENDING = "client-asset-pending",
+  CLIENT_ASSET_BROKEN = "client-asset-broken",
+  CLIENT_ASSET_WAITING_CONFIRM = "client-asset-waitingConfirm",
+  CLIENT_ASSETS_EXPIRES = "client-assets-expires",
+  WEBHOOK = "webhook",
+  KOMU_LOGS = "komu_logs",
+  WEBHOOK_LOGS = "webhook_logs",
+  ASSET_RENTAL_CUSTOMERS = "asset-rental-customers",
+}
+
+// Resource labels enum
+export enum RESOURCE_LABELS {
+  USERS = "users",
+  ASSETS = "assets",
+  TOOLS = "tools",
+  CLIENT_ASSETS = "client-assets",
+  CONSUMABLES = "consumables",
+  TAX_TOKEN = "tax_token",
+}
+
+// Resource translation keys enum
+export enum RESOURCE_TRANSLATION_KEYS {
+  USERS = "resource.users",
+  DASHBOARD = "resource.dashboard",
+  DETAIL_DEVICE = "resource.detail-device",
+  CHECKIN_CHECKOUT = "resource.checkin-checkout",
+  ASSETS = "resource.assets",
+  ASSETS_ASSIGN = "resource.assets-assign",
+  ASSETS_READY_TO_DEPLOY = "resource.assets-readyToDeploy",
+  ASSETS_PENDING = "resource.assets-pending",
+  ASSETS_BROKEN = "resource.assets-broken",
+  ASSETS_MAINTENANCE = "resource.assets-maintenance",
+  TOOLS_ALL = "resource.tools-all",
+  TOOLS_WAITING = "resource.tools-waiting",
+  LICENSES = "resource.licenses",
+  USERS_TOOLS = "resource.users-tools",
+  USERS_TAX_TOKENS = "resource.users-tax-tokens",
+  USERS_LICENSES = "resource.users_licenses",
+  MODELS = "resource.model",
+  CATEGORY = "resource.category",
+  MANUFACTURES = "resource.manufactures",
+  SUPPLIERS = "resource.suppliers",
+  DEPARTMENT = "resource.department",
+  LOCATION = "resource.location",
+  TAX_TOKEN_ASSIGN = "resource.tax_token_assign",
+  REQUEST = "resource.request",
+  REPORT = "resource.report",
+  MANAGER_USER = "resource.manager_user",
+  ASSETS_WAITING_CONFIRM = "resource.assets-waiting-confirm",
+  CONSUMABLES = "resource.consumables",
+  CONSUMABLES_MAINTENANCE = "resource.consumables-maintenance",
+  ACCESSORY = "resource.accessory",
+  ACCESSORY_DETAILS = "resource.accessory_details",
+  CONSUMABLE_DETAILS = "resource.consumable_details",
+  SUPPLIER_DETAILS = "resource.supplier_details",
+  LOCATION_DETAILS = "resource.location_details",
+  MANUFACTURES_DETAILS = "resource.manufactures_details",
+  WEBHOOK_DETAILS = "resource.webhook_details",
+  ASSETS_EXPIRES = "resource.assets-expires",
+  TAX_TOKEN = "resource.tax_token",
+  TAX_TOKEN_WAITING = "resource.tax_token_waiting",
+  CLIENT_ASSETS = "resource.client-assets",
+  CLIENT_ASSET_ASSIGNED = "resource.client-asset-assigned",
+  CLIENT_ASSET_READY_TO_DEPLOY = "resource.client-asset-readyToDeploy",
+  CLIENT_ASSET_PENDING = "resource.client-asset-pending",
+  CLIENT_ASSET_BROKEN = "resource.client-asset-broken",
+  CLIENT_ASSET_WAITING_CONFIRM = "resource.client-asset-waitingConfirm",
+  WEBHOOK = "resource.webhook",
+  KOMU_LOGS = "resource.komu_logs",
+  WEBHOOK_LOGS = "resource.webhook_logs",
+  ASSET_RENTAL_CUSTOMERS = "resource.asset-rental-customers",
+}
